Fix Issue 41: Tabs stay open when connection is closed. Fix Issue 49: Titles of connection-bound tabs should be of the format [Tab Name] - [Connection Name]
Refactored IConnectionBoundCtl to use a ConnectionInfo property instead of IConnection. As a result, many changes from IConnection to ConnectionInfo were required throughout the entire codebase.

This allows us to keep track of how many open instances there are by the connection name. Thus making it easy for us if we need to warn the user of any open tabs. 

A positive consequence of this change is that it allows us to easily fix Issue 49 as well as we can now infer the Tab's title text from the refactored property. 

Another positive consequence is that the Connection-bound controls are now aware of connection renaming.