Fix Issue 55: Coordinate System Browser/Catalog
Added new Thirdparty component: SQLite ADO.net driver

Added Coordinate System Manager control and Coordinate System Picker dialog

Spatial Context Information Dialog now has an extra button "Load CS" which invokes the Coordinate System Picker. If a CS is chosen, the name, description, cs and wkt fields of the spatial context are filled in automatically (the name field won't be filled in if it is disabled)

Coordinate System Definitions are stored in the cscatalog.sqlite database. 

This is a SQLite v3 database with a single table "Projections" with the following definition:
- Name : TEXT, Primary Key
- Description : TEXT, NULL
- WKT : TEXT, NOT NULL