Fix Issue 64: Support "external" FDO install
New string preference: pref_str_fdo_home (Default value: [Fdo Toolbox path]\FDO] )
If this path does not exist, the user will be prompted for the path to the FDO libraries. 
pref_str_fdo_home will be set to this path and this path will be used for subsequent sessions.