Fix Issue 101: Bulk Copy progress may exceed 100%
The issue is with the Count() select aggregate function. It used the first property in the source class property list. This list does not include identity properties, thus there is a high probability of a property count being less than the actual total.

The code was modified to use Count() with the first identity property in the source class definition. As identity properties cannot be null, the count is guaranteed to be the total.